Fix an issue with very small periods and large time requested generating negative times on overflow.  This fix will resolve it for the time being but some changes are needed in the timeRepresentation to completely resolve it.  This fix just makes it significantly harder to generate the bug.  like (2^24X harder) in numerical terms.

+++ src/helics/core/TimeCoordinator.cpp
@@ -310,7 +310,7 @@ Time TimeCoordinator::generateAllowedTime (Time testTime) const
         }
         if (testTime - time_grantBase > info.period)
         {
-            auto blk = static_cast<int> (std::ceil ((testTime - time_grantBase) / info.period));
+            auto blk = std::ceil ((testTime - time_grantBase) / info.period);
             testTime = time_grantBase + blk * info.period;
         }
         else
